api/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
python-dotenv
pydantic[email]
